refactor: 에러 메시지 상수 선언
